requirements.txt
Django>=3.2
psycopg2-binary>=2.8
Pillow>=2.6.1
mysqlclient>=1.4.2
python-decouple>=3.1
django-model-utils
djangorestframework
django-rosetta
django-parler>=2.3

